+++ src/main/java/moze_intel/projecte/gameObjs/entity/EntityLensProjectile.java
@@ -13,20 +13,20 @@ import net.minecraft.world.WorldServer;
 
 public class EntityLensProjectile extends PEProjectile
 {
-	private byte charge;
+	private int charge;
 	
 	public EntityLensProjectile(World world) 
 	{
 		super(world);
 	}
 
-	public EntityLensProjectile(World world, EntityPlayer entity, byte charge)
+	public EntityLensProjectile(World world, EntityPlayer entity, int charge)
 	{
 		super(world, entity);
 		this.charge = charge;
 	}
 
-	public EntityLensProjectile(World world, double x, double y, double z, byte charge) 
+	public EntityLensProjectile(World world, double x, double y, double z, int charge)
 	{
 		super(world, x, y, z);
 		this.charge = charge;
@@ -66,12 +66,12 @@ public class EntityLensProjectile extends PEProjectile
 	public void writeEntityToNBT(NBTTagCompound nbt)
 	{
 		super.writeEntityToNBT(nbt);
-		nbt.setByte("Charge", charge);
+		nbt.setInteger("Charge", charge);
 	}
 	
 	public void readEntityFromNBT(NBTTagCompound nbt)
 	{
 		super.readEntityFromNBT(nbt);
-		charge = nbt.getByte("Charge");
+		charge = nbt.getInteger("Charge");
 	}
 }
